test: add consolidated Passage directive suite
