+++ apps/campfire/src/components/Passage/__tests__/Passage.directives.test.tsx
@@ -1,4 +1,4 @@
-import { describe, it, expect, beforeEach } from 'bun:test'
+import { describe, it, beforeEach, expect } from 'bun:test'
 import {
   render,
   screen,
@@ -14,6 +14,180 @@ import { useStoryDataStore } from '@campfire/state/useStoryDataStore'
 import { useGameStore } from '@campfire/state/useGameStore'
 import { resetStores } from '@campfire/test-utils/helpers'
 
+/**
+ * Consolidated tests for various Passage directives.
+ */
+
+/**
+ * Verifies that multiple image directives render each image on a slide.
+ */
+describe('Passage image directive', () => {
+  beforeEach(async () => {
+    document.body.innerHTML = ''
+    resetStores()
+    ;(HTMLElement.prototype as any).animate = () => ({
+      finished: Promise.resolve(),
+      cancel() {},
+      finish() {}
+    })
+    if (!i18next.isInitialized) {
+      await i18next.use(initReactI18next).init({ lng: 'en-US', resources: {} })
+    } else {
+      await i18next.changeLanguage('en-US')
+      i18next.services.resourceStore.data = {}
+    }
+  })
+
+  it('renders multiple SlideImage components', async () => {
+    const passage: Element = {
+      type: 'element',
+      tagName: 'tw-passagedata',
+      properties: { pid: '1', name: 'Start' },
+      children: [
+        {
+          type: 'text',
+          value: `:::deck{size='1280x720' hideNavigation=true}\n            :::slide\n    :::reveal{at=0}\n      ::image{src='https://placecats.com/bella/1280/360'}\n\n      ::image{src='https://placecats.com/neo/250/250' x=50 y=250 className='rounded-full shadow-lg'}\n    :::\n  :::\n:::`
+        }
+      ]
+    }
+    useStoryDataStore.setState({ passages: [passage], currentPassageId: '1' })
+    render(<Passage />)
+    const images = await screen.findAllByTestId('slideImage')
+    expect(images).toHaveLength(2)
+    expect(images[0].querySelector('img')?.getAttribute('src')).toBe(
+      'https://placecats.com/bella/1280/360'
+    )
+    expect(images[1].querySelector('img')?.getAttribute('src')).toBe(
+      'https://placecats.com/neo/250/250'
+    )
+  })
+})
+
+/**
+ * Verifies that the layer directive renders the Layer component in a Passage.
+ */
+describe('Passage layer directive', () => {
+  beforeEach(async () => {
+    document.body.innerHTML = ''
+    resetStores()
+    if (!i18next.isInitialized) {
+      await i18next.use(initReactI18next).init({ lng: 'en-US', resources: {} })
+    } else {
+      await i18next.changeLanguage('en-US')
+      i18next.services.resourceStore.data = {}
+    }
+  })
+
+  it('renders a Layer component', async () => {
+    const passage: Element = {
+      type: 'element',
+      tagName: 'tw-passagedata',
+      properties: { pid: '1', name: 'Start' },
+      children: [{ type: 'text', value: ':::layer{x=10}\nContent\n:::' }]
+    }
+
+    useStoryDataStore.setState({ passages: [passage], currentPassageId: '1' })
+    render(<Passage />)
+
+    const el = await screen.findByTestId('layer')
+    expect(el.tagName).toBe('DIV')
+    expect(el.textContent).toContain('Content')
+  })
+})
+
+/**
+ * Tests rendering of the shape directive within a Passage.
+ */
+describe('Passage shape directive', () => {
+  beforeEach(async () => {
+    document.body.innerHTML = ''
+    resetStores()
+    const animation: Partial<Animation> = {
+      finished: Promise.resolve<Animation>({} as Animation),
+      cancel() {}
+    }
+    Object.defineProperty(HTMLElement.prototype, 'animate', {
+      value: () => animation as Animation
+    })
+    if (!i18next.isInitialized) {
+      await i18next.use(initReactI18next).init({ lng: 'en-US', resources: {} })
+    } else {
+      await i18next.changeLanguage('en-US')
+      i18next.services.resourceStore.data = {}
+    }
+  })
+
+  it('renders SlideShape components without stray markers', async () => {
+    const passage: Element = {
+      type: 'element',
+      tagName: 'tw-passagedata',
+      properties: { pid: '1', name: 'Start' },
+      children: [
+        {
+          type: 'text',
+          value:
+            ':::deck{size=800x600}\n:::slide\n::shape{x=10 y=20 w=100 h=50 type="rect" data-test="ok"}\n:::\n:::\n'
+        }
+      ]
+    }
+    useStoryDataStore.setState({ passages: [passage], currentPassageId: '1' })
+    render(<Passage />)
+    const el = await screen.findByTestId('slideShape')
+    expect(el).toBeTruthy()
+    expect(el.style.left).toBe('10px')
+    expect(el.style.top).toBe('20px')
+    expect(el.style.width).toBe('100px')
+    expect(el.style.height).toBe('50px')
+    expect(el.getAttribute('data-test')).toBe('ok')
+    expect(document.body.innerHTML).not.toContain('<SlideShape')
+    expect(document.body.textContent).not.toContain(':::')
+  })
+})
+
+/**
+ * Tests rendering of the text directive within a Passage.
+ */
+describe('Passage text directive', () => {
+  beforeEach(async () => {
+    document.body.innerHTML = ''
+    resetStores()
+    ;(HTMLElement.prototype as any).animate = () => ({
+      finished: Promise.resolve(),
+      cancel() {}
+    })
+    if (!i18next.isInitialized) {
+      await i18next.use(initReactI18next).init({ lng: 'en-US', resources: {} })
+    } else {
+      await i18next.changeLanguage('en-US')
+      i18next.services.resourceStore.data = {}
+    }
+  })
+
+  it('renders SlideText components without stray markers', async () => {
+    const passage: Element = {
+      type: 'element',
+      tagName: 'tw-passagedata',
+      properties: { pid: '1', name: 'Start' },
+      children: [
+        {
+          type: 'text',
+          value:
+            ':::deck{size=800x600}\n:::slide\n:::text{x=80 y=80 as="h2"}\nHello\n:::\n:::\n:::\n'
+        }
+      ]
+    }
+    useStoryDataStore.setState({ passages: [passage], currentPassageId: '1' })
+    render(<Passage />)
+    const el = await screen.findByTestId('slideText')
+    expect(el).toBeTruthy()
+    expect(document.body.innerHTML).not.toContain('<SlideText')
+    expect(document.body.textContent).not.toContain(':::')
+  })
+})
+
+/**
+ * Tests for trigger directives in Passage.
+ */
 describe('Passage trigger directives', () => {
   beforeEach(async () => {
     document.body.innerHTML = ''
@@ -233,14 +407,12 @@ describe('Passage trigger directives', () => {
     useStoryDataStore.setState({ passages: [passage], currentPassageId: '1' })
     render(<Passage />)
     const button = await screen.findByRole('button', { name: 'Styled Label' })
-    // Ensure wrapper element is used inside the button
     const wrapper = button.querySelector(
       '[data-testid="wrapper"]'
     ) as HTMLElement
     expect(wrapper).toBeTruthy()
     expect(wrapper.className).toContain('campfire-wrapper')
     expect(wrapper.className).toContain('fancy')
-    // Clicking runs remaining directive content
     act(() => {
       button.click()
     })
@@ -259,7 +431,6 @@ describe('Passage trigger directives', () => {
           type: 'text',
           value:
             ':::trigger{label="Fallback"}\n' +
-            // Two wrappers (only one should be used)
             ':::wrapper{as="p" className="first"}\n' +
             'First\n' +
             ':::\n' +
@@ -275,7 +446,6 @@ describe('Passage trigger directives', () => {
     const button = await screen.findByRole('button', { name: 'First' })
     const wrappers = button.querySelectorAll('[data-testid="wrapper"]')
     expect(wrappers).toHaveLength(1)
-    // Coerced to span when used as trigger label
     expect((wrappers[0] as HTMLElement).tagName).toBe('SPAN')
   })
